backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-tenants==3.7.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
celery==5.4.0
redis==5.1.1
django-redis==5.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
requests==2.32.3
resend==2.4.0
flower==2.0.1
gunicorn==23.0.0
whitenoise==6.7.0
django-filter==24.3
drf-spectacular==0.27.2
